requirements.txt
flask==2.0.1
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.0.2

opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
albumentations==1.3.1
pandas==2.0.3
torch==2.1.2
torchvision==0.16.2
wandb==0.16.1
numpy==1.24.3
scipy==1.11.3
PyYAML==6.0.1
scikit-learn==1.3.2
tqdm==4.66.1
pillow==10.1.0
timm==0.9.12
jupyterlab==4.0.9
tabulate==0.9.0

# URL to the custom wheel file
https://github.com/BohemianVRA/FGVC/releases/download/v1.5.4-dev/fgvc-1.5.4.dev0-py3-none-any.whl



